@jreback 2021-09-05
also since this is a huge speedup this migth improve some concat / join asvs for multi-multi (if so that is worth a whatsnew note0